actualizacion de servicios

+++ controllers/userController.js
@@ -4,11 +4,13 @@ const sendEmail = require('../utils/email');
 
 const createUser = async (req, res) => {
     try {
-        const { body } = req;
+        const { body, tenantId } = req;
 
         // create user
-        const user = new User(body);
+        body.tenantId = tenantId;
+        body.password = Math.random().toString(36).slice(-8);
 
+        const user = new User(body);
         await user.save();
 
         res.status(201).json({ status: 'ok', message: 'User created successfully', data: user });
@@ -19,7 +21,9 @@ const createUser = async (req, res) => {
 
 const getUsers = async (req, res) => {
     try {
-        const users = await User.find().populate({ path: 'companyId', as: 'Company' });
+        const { tenantId } = req;
+
+        const users = await User.find({ tenantId }).populate({ path: 'urbanizationId', as: 'Urbanization' });
         res.status(200).json({ status: 'ok', message: 'Users found successfully', data: users });
     } catch (err) {
         res.status(400).json({ status: 'error', message: `Error find user: ${err.message}`, data: null });
@@ -29,20 +33,10 @@ const getUsers = async (req, res) => {
 const getUser = async (req, res) => {
     try {
         const { _id } = req.params;
+        const { tenantId } = req;
 
-        const user = await User.findById({ _id });
+        const user = await User.findById({ _id, tenantId });
         res.status(201).json({ status: 'ok', message: 'User find successfully', data: user });
-    } catch (err) {
-        res.status(400).json({ status: 'error', message: `Error find user: ${err.message}`, data: null });
-    }
-};
-
-const getUserByEmail = async (req, res) => {
-    try {
-        const { email } = req.params;
-
-        const user = await User.findOne({ email });
-        res.status(201).json({ status: "ok", message: 'User find successfully', data: user });
     } catch (err) {
         res.status(400).json({ status: 'error', message: `Error find user: ${err.message}`, data: null });
     }
@@ -51,9 +45,9 @@ const getUserByEmail = async (req, res) => {
 const updateUser = async (req, res) => {
     try {
         const { _id } = req.params;
-        const { body } = req;
+        const { body, tenantId } = req;
 
-        const user = await User.findByIdAndUpdate(_id, body, { new: true });
+        const user = await User.findByIdAndUpdate({ _id, tenantId }, body, { new: true });
         await user.save();
 
         res.status(201).json({ status: 'ok', message: 'Profile update successfully', data: user });
@@ -66,8 +60,9 @@ const updateProfile = async (req, res) => {
     try {
         const { _id } = req.params;
         const { name, email } = req.body;
+        const { tenantId } = req;
 
-        const user = await User.findById({ _id });
+        const user = await User.findById({ _id, tenantId });
 
         // updating user
         if (name) user.name = name;
@@ -85,8 +80,9 @@ const changePassword = async (req, res) => {
     try {
         const { _id } = req.params;
         const { oldPassword, newPassword } = req.body;
+        const { tenantId } = req;
 
-        const user = await User.findById({ _id });
+        const user = await User.findById({ _id, tenantId });
 
         if (user && !(await bcrypt.compare(oldPassword, user.password))) {
             return res.status(400).json({ message: 'La contraseña ingresada es incorrecta' });
@@ -107,8 +103,9 @@ const changePassword = async (req, res) => {
 const deleteUser = async (req, res) => {
     try {
         const { _id } = req.params;
+        const { tenantId } = req;
 
-        await User.findByIdAndDelete(_id);
+        await User.findByIdAndDelete({ _id, tenantId });
 
         res.status(200).json({ status: 'ok', message: 'User deleted successfully', data: null });
     } catch (err) {
@@ -148,5 +145,16 @@ const resetPassword = async (req, res) => {
         res.status(400).json({ status: 'error', message: `Error restableciendo la contraseña: ${error.message}`, data: null });
     }
 }
+
+const getUserByEmail = async (req, res) => {
+    try {
+        const { email } = req.params;
+
+        const user = await User.findOne({ email });
+        res.status(201).json({ status: "ok", message: 'User find successfully', data: user });
+    } catch (err) {
+        res.status(400).json({ status: 'error', message: `Error find user: ${err.message}`, data: null });
+    }
+};
 
 module.exports = { createUser, getUsers, getUser, getUserByEmail, updateUser, updateProfile, changePassword, deleteUser, resetPassword };
